pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dorsal-whisper"
version = "0.4.0"
description = "Dorsal model: dorsal-whisper"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dorsalhub>=0.8.6",
    "faster-whisper>=0.10.0",
    "protobuf>=6.33.5",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
    "mypy"
]

[project.entry-points."dorsal.models"]
dorsal_whisper = "dorsal_whisper"

[tool.hatch.build.targets.wheel]
packages = ["dorsal_whisper"]

[tool.hatch.build.targets.wheel.force-include]
"model_config.toml" = "dorsal_whisper/model_config.toml"
